side bar prototype
