+++ src/styles/views-styles/galatic-view.module.css
@@ -1,7 +1,6 @@
 .galaxy {
   height: 100vh;
   width: 100vw;
-  position:absolute;
   z-index: -99;
 }
 .sectora {
